+++ components/hord-cli/src/core/protocol/numbering.rs
@@ -1,284 +1,306 @@
-use chainhook_sdk::types::{BitcoinBlockData, OrdinalOperation, TransactionIdentifier};
+use chainhook_sdk::types::{BlockIdentifier, TransactionIdentifier};
 use chainhook_sdk::utils::Context;
 use dashmap::DashMap;
 use fxhash::FxHasher;
-use rand::seq::SliceRandom;
-use rand::thread_rng;
-use rusqlite::Connection;
-use std::collections::{HashMap, VecDeque};
 use std::hash::BuildHasherDefault;
-use std::sync::mpsc::channel;
+use std::path::PathBuf;
 use std::sync::Arc;
 
-use crate::core::HordConfig;
-use crate::db::{find_all_inscriptions_in_block, retrieve_satoshi_point_using_lazy_storage_v3};
+use crate::db::{
+    find_lazy_block_at_block_height, open_readonly_hord_db_conn_rocks_db_loop, TransferData,
+};
 
 use crate::db::{LazyBlockTransaction, TraversalResult};
+use crate::ord::height::Height;
 
-pub fn retrieve_inscribed_satoshi_points_from_block_v3(
-    block: &BitcoinBlockData,
-    next_blocks: &Vec<BitcoinBlockData>,
-    cache_l1: &mut HashMap<(TransactionIdentifier, usize), TraversalResult>,
-    cache_l2: &Arc<DashMap<(u32, [u8; 8]), LazyBlockTransaction, BuildHasherDefault<FxHasher>>>,
-    inscriptions_db_conn: &mut Connection,
-    hord_config: &HordConfig,
+pub fn retrieve_satoshi_point_using_lazy_storage_v3(
+    blocks_db_dir: &PathBuf,
+    block_identifier: &BlockIdentifier,
+    transaction_identifier: &TransactionIdentifier,
+    inscription_input_index: usize,
+    inscription_number: i64,
+    traversals_cache: &Arc<
+        DashMap<(u32, [u8; 8]), LazyBlockTransaction, BuildHasherDefault<FxHasher>>,
+    >,
     ctx: &Context,
-) -> Result<bool, String> {
-    let (mut transactions_ids, l1_cache_hits) =
-        get_transactions_to_process(block, cache_l1, inscriptions_db_conn, ctx);
+) -> Result<TraversalResult, String> {
+    let mut inscription_offset_intra_output = 0;
+    let mut inscription_output_index: usize = 0;
+    let mut ordinal_offset = 0;
+    let mut ordinal_block_number = block_identifier.index as u32;
+    let txid = transaction_identifier.get_8_hash_bytes();
 
-    let inner_ctx = if hord_config.logs.ordinals_computation {
-        ctx.clone()
-    } else {
-        Context::empty()
-    };
+    let mut blocks_db = open_readonly_hord_db_conn_rocks_db_loop(&blocks_db_dir, &ctx);
 
-    let has_transactions_to_process = !transactions_ids.is_empty() || !l1_cache_hits.is_empty();
-
-    let thread_max = hord_config.ingestion_thread_max * 2;
-
-    if has_transactions_to_process {
-        let expected_traversals = transactions_ids.len() + l1_cache_hits.len();
-        let (traversal_tx, traversal_rx) = channel();
-
-        let mut tx_thread_pool = vec![];
-        let mut thread_pool_handles = vec![];
-
-        for thread_index in 0..thread_max {
-            let (tx, rx) = channel();
-            tx_thread_pool.push(tx);
-
-            let moved_traversal_tx = traversal_tx.clone();
-            let moved_ctx = inner_ctx.clone();
-            let moved_hord_db_path = hord_config.db_path.clone();
-            let local_cache = cache_l2.clone();
-
-            let handle = hiro_system_kit::thread_named("Worker")
-                .spawn(move || {
-                    while let Ok(Some((
-                        transaction_id,
-                        block_identifier,
-                        input_index,
-                        prioritary,
-                    ))) = rx.recv()
-                    {
-                        let traversal: Result<TraversalResult, String> =
-                            retrieve_satoshi_point_using_lazy_storage_v3(
-                                &moved_hord_db_path,
-                                &block_identifier,
-                                &transaction_id,
-                                input_index,
-                                0,
-                                &local_cache,
-                                &moved_ctx,
-                            );
-                        let _ = moved_traversal_tx.send((traversal, prioritary, thread_index));
-                    }
-                })
-                .expect("unable to spawn thread");
-            thread_pool_handles.push(handle);
+    let (sats_ranges, inscription_offset_cross_outputs) = match traversals_cache
+        .get(&(block_identifier.index as u32, txid.clone()))
+    {
+        Some(entry) => {
+            let tx = entry.value();
+            (
+                tx.get_sat_ranges(),
+                tx.get_cumulated_sats_in_until_input_index(inscription_input_index),
+            )
         }
-
-        // Empty cache
-        let mut thread_index = 0;
-        for key in l1_cache_hits.iter() {
-            if let Some(entry) = cache_l1.remove(key) {
-                let _ = traversal_tx.send((Ok(entry), true, thread_index));
-                thread_index = (thread_index + 1) % thread_max;
+        None => {
+            let mut attempt = 0;
+            loop {
+                match find_lazy_block_at_block_height(
+                    ordinal_block_number,
+                    3,
+                    false,
+                    &blocks_db,
+                    &ctx,
+                ) {
+                    None => {
+                        if attempt < 3 {
+                            attempt += 1;
+                            blocks_db =
+                                open_readonly_hord_db_conn_rocks_db_loop(&blocks_db_dir, &ctx);
+                        } else {
+                            return Err(format!("block #{ordinal_block_number} not in database"));
+                        }
+                    }
+                    Some(block) => match block.find_and_serialize_transaction_with_txid(&txid) {
+                        Some(tx) => {
+                            let sats_ranges = tx.get_sat_ranges();
+                            let inscription_offset_cross_outputs =
+                                tx.get_cumulated_sats_in_until_input_index(inscription_input_index);
+                            traversals_cache.insert((ordinal_block_number, txid.clone()), tx);
+                            break (sats_ranges, inscription_offset_cross_outputs);
+                        }
+                        None => return Err(format!("txid not in block #{ordinal_block_number}")),
+                    },
+                }
             }
         }
+    };
 
-        ctx.try_log(|logger| {
-            info!(
-                logger,
-                "Number of inscriptions in block #{} to process: {} (L1 cache hits: {}, queue len: {}, L1 cache len: {}, L2 cache len: {})",
-                block.block_identifier.index,
-                transactions_ids.len(),
-                l1_cache_hits.len(),
-                next_blocks.len(),
-                cache_l1.len(),
-                cache_l2.len(),
-            )
-        });
+    for (i, (min, max)) in sats_ranges.into_iter().enumerate() {
+        if inscription_offset_cross_outputs >= min && inscription_offset_cross_outputs < max {
+            inscription_output_index = i;
+            inscription_offset_intra_output = inscription_offset_cross_outputs - min;
+        }
+    }
+    ctx.try_log(|logger| {
+        info!(
+            logger,
+            "Computing ordinal number for Satoshi point {} ({}:0 -> {}:{}/{})  (block #{})",
+            transaction_identifier.hash,
+            inscription_input_index,
+            inscription_output_index,
+            inscription_offset_intra_output,
+            inscription_offset_cross_outputs,
+            block_identifier.index
+        )
+    });
 
-        let mut rng = thread_rng();
-        transactions_ids.shuffle(&mut rng);
-        let mut priority_queue = VecDeque::new();
-        let mut warmup_queue = VecDeque::new();
+    let mut tx_cursor: ([u8; 8], usize) = (txid, inscription_input_index);
+    let mut hops: u32 = 0;
 
-        for (transaction_id, input_index) in transactions_ids.into_iter() {
-            priority_queue.push_back((
-                transaction_id,
-                block.block_identifier.clone(),
-                input_index,
-                true,
+    loop {
+        hops += 1;
+        if hops as u64 > block_identifier.index {
+            return Err(format!(
+                "Unable to process transaction {} detected after {hops} iterations. Manual investigation required",
+                transaction_identifier.hash
             ));
         }
 
-        // Feed each workers with 2 workitems each
-        for thread_index in 0..thread_max {
-            let _ = tx_thread_pool[thread_index].send(priority_queue.pop_front());
-        }
-        for thread_index in 0..thread_max {
-            let _ = tx_thread_pool[thread_index].send(priority_queue.pop_front());
+        if let Some(cached_tx) = traversals_cache.get(&(ordinal_block_number, tx_cursor.0)) {
+            let tx = cached_tx.value();
+
+            let mut next_found_in_cache = false;
+            let mut sats_out = 0;
+            for (index, output_value) in tx.outputs.iter().enumerate() {
+                if index == tx_cursor.1 {
+                    break;
+                }
+                sats_out += output_value;
+            }
+            sats_out += ordinal_offset;
+
+            let mut sats_in = 0;
+            for input in tx.inputs.iter() {
+                sats_in += input.txin_value;
+
+                if sats_out < sats_in {
+                    ordinal_offset = sats_out - (sats_in - input.txin_value);
+                    ordinal_block_number = input.block_height;
+                    tx_cursor = (input.txin.clone(), input.vout as usize);
+                    next_found_in_cache = true;
+                    break;
+                }
+            }
+
+            if next_found_in_cache {
+                continue;
+            }
+
+            if sats_in == 0 {
+                ctx.try_log(|logger| {
+                    error!(
+                        logger,
+                        "Transaction {} is originating from a non spending transaction",
+                        transaction_identifier.hash
+                    )
+                });
+                return Ok(TraversalResult {
+                    inscription_number: 0,
+                    ordinal_number: 0,
+                    transfers: 0,
+                    inscription_input_index,
+                    transaction_identifier_inscription: transaction_identifier.clone(),
+                    transfer_data: TransferData {
+                        inscription_offset_intra_output,
+                        transaction_identifier_location: transaction_identifier.clone(),
+                        output_index: inscription_output_index,
+                        tx_index: 0,
+                    },
+                });
+            }
         }
 
-        let mut next_block_iter = next_blocks.iter();
-        let mut traversals_received = 0;
-        while let Ok((traversal_result, prioritary, thread_index)) = traversal_rx.recv() {
-            if prioritary {
-                traversals_received += 1;
-            }
-            match traversal_result {
-                Ok(traversal) => {
-                    inner_ctx.try_log(|logger| {
-                        info!(
-                            logger,
-                            "Satoshi #{} was minted in block #{} at offset {} and was transferred {} times (progress: {traversals_received}/{expected_traversals}) (priority queue: {prioritary}, thread: {thread_index}).",
-                            traversal.ordinal_number, traversal.get_ordinal_coinbase_height(), traversal.get_ordinal_coinbase_offset(), traversal.transfers
-                            )
-                    });
-                    cache_l1.insert(
-                        (
-                            traversal.transaction_identifier_inscription.clone(),
-                            traversal.inscription_input_index,
-                        ),
-                        traversal,
-                    );
-                }
-                Err(e) => {
-                    ctx.try_log(|logger| {
-                        error!(logger, "Unable to compute inscription's Satoshi: {e}",)
-                    });
+        let lazy_block = {
+            let mut attempt = 0;
+            loop {
+                match find_lazy_block_at_block_height(
+                    ordinal_block_number,
+                    3,
+                    false,
+                    &blocks_db,
+                    &ctx,
+                ) {
+                    Some(block) => break block,
+                    None => {
+                        if attempt < 3 {
+                            attempt += 1;
+                            blocks_db =
+                                open_readonly_hord_db_conn_rocks_db_loop(&blocks_db_dir, &ctx);
+                        } else {
+                            return Err(format!("block #{ordinal_block_number} not in database"));
+                        }
+                    }
                 }
             }
-            if traversals_received == expected_traversals {
+        };
+
+        let coinbase_txid = lazy_block.get_coinbase_txid();
+        let txid = tx_cursor.0;
+
+        // evaluate exit condition: did we reach the **final** coinbase transaction
+        if coinbase_txid.eq(&txid) {
+            let subsidy = Height(ordinal_block_number.into()).subsidy();
+            if ordinal_offset < subsidy {
+                // Great!
                 break;
             }
 
-            if let Some(w) = priority_queue.pop_front() {
-                let _ = tx_thread_pool[thread_index].send(Some(w));
-            } else {
-                if let Some(w) = warmup_queue.pop_front() {
-                    let _ = tx_thread_pool[thread_index].send(Some(w));
-                } else {
-                    if let Some(next_block) = next_block_iter.next() {
-                        let (mut transactions_ids, _) = get_transactions_to_process(
-                            next_block,
-                            cache_l1,
-                            inscriptions_db_conn,
-                            ctx,
-                        );
+            // loop over the transaction fees to detect the right range
+            let mut accumulated_fees = subsidy;
 
-                        ctx.try_log(|logger| {
-                            info!(
-                                logger,
-                                "Number of inscriptions in block #{} to pre-process: {}",
-                                block.block_identifier.index,
-                                transactions_ids.len()
-                            )
-                        });
+            for tx in lazy_block.iter_tx() {
+                let mut total_in = 0;
+                for input in tx.inputs.iter() {
+                    total_in += input.txin_value;
+                }
 
-                        transactions_ids.shuffle(&mut rng);
-                        for (transaction_id, input_index) in transactions_ids.into_iter() {
-                            warmup_queue.push_back((
-                                transaction_id,
-                                next_block.block_identifier.clone(),
-                                input_index,
-                                false,
-                            ));
+                let mut total_out = 0;
+                for output_value in tx.outputs.iter() {
+                    total_out += output_value;
+                }
+
+                let fee = total_in - total_out;
+                if accumulated_fees + fee > ordinal_offset {
+                    // We are looking at the right transaction
+                    // Retraverse the inputs to select the index to be picked
+                    let offset_within_fee = ordinal_offset - accumulated_fees;
+                    total_out += offset_within_fee;
+                    let mut sats_in = 0;
+
+                    for input in tx.inputs.into_iter() {
+                        sats_in += input.txin_value;
+
+                        if sats_in > total_out {
+                            ordinal_offset = total_out - (sats_in - input.txin_value);
+                            ordinal_block_number = input.block_height;
+                            tx_cursor = (input.txin.clone(), input.vout as usize);
+                            break;
                         }
-                        let _ = tx_thread_pool[thread_index].send(warmup_queue.pop_front());
                     }
+                    break;
+                } else {
+                    accumulated_fees += fee;
                 }
             }
-        }
-
-        for tx in tx_thread_pool.iter() {
-            // Empty the queue
-            if let Ok((traversal_result, prioritary, thread_index)) = traversal_rx.try_recv() {
-                if let Ok(traversal) = traversal_result {
-                    inner_ctx.try_log(|logger| {
-                            info!(
-                                logger,
-                                "Satoshi #{} was minted in block #{} at offset {} and was transferred {} times (progress: {traversals_received}/{expected_traversals}) (priority queue: {prioritary}, thread: {thread_index}).",
-                                traversal.ordinal_number, traversal.get_ordinal_coinbase_height(), traversal.get_ordinal_coinbase_offset(), traversal.transfers
-                                )
-                        });
-                    cache_l1.insert(
-                        (
-                            traversal.transaction_identifier_inscription.clone(),
-                            traversal.inscription_input_index,
-                        ),
-                        traversal,
-                    );
-                }
-            }
-            let _ = tx.send(None);
-        }
-
-        let _ = hiro_system_kit::thread_named("Garbage collection").spawn(move || {
-            for handle in thread_pool_handles.into_iter() {
-                let _ = handle.join();
-            }
-        });
-    } else {
-        ctx.try_log(|logger| {
-            info!(
-                logger,
-                "No inscriptions to index in block #{}", block.block_identifier.index
-            )
-        });
-    }
-    Ok(has_transactions_to_process)
-}
-
-fn get_transactions_to_process(
-    block: &BitcoinBlockData,
-    cache_l1: &mut HashMap<(TransactionIdentifier, usize), TraversalResult>,
-    inscriptions_db_conn: &mut Connection,
-    ctx: &Context,
-) -> (
-    Vec<(TransactionIdentifier, usize)>,
-    Vec<(TransactionIdentifier, usize)>,
-) {
-    let mut transactions_ids: Vec<(TransactionIdentifier, usize)> = vec![];
-    let mut l1_cache_hits = vec![];
-
-    let mut known_transactions =
-        find_all_inscriptions_in_block(&block.block_identifier.index, inscriptions_db_conn, ctx);
-
-    for tx in block.transactions.iter().skip(1) {
-        // Have a new inscription been revealed, if so, are looking at a re-inscription
-        for ordinal_event in tx.metadata.ordinal_operations.iter() {
-            let inscription_data = match ordinal_event {
-                OrdinalOperation::InscriptionRevealed(inscription_data) => inscription_data,
-                OrdinalOperation::CursedInscriptionRevealed(inscription_data) => inscription_data,
-                OrdinalOperation::InscriptionTransferred(_) => {
-                    continue;
-                }
+        } else {
+            // isolate the target transaction
+            let lazy_tx = match lazy_block.find_and_serialize_transaction_with_txid(&txid) {
+                Some(entry) => entry,
+                None => unreachable!(),
             };
-            let key = (
-                tx.transaction_identifier.clone(),
-                inscription_data.inscription_input_index,
-            );
-            if cache_l1.contains_key(&key) {
-                l1_cache_hits.push(key);
-                continue;
+
+            let mut sats_out = 0;
+            for (index, output_value) in lazy_tx.outputs.iter().enumerate() {
+                if index == tx_cursor.1 {
+                    break;
+                }
+                sats_out += output_value;
+            }
+            sats_out += ordinal_offset;
+
+            let mut sats_in = 0;
+            for input in lazy_tx.inputs.iter() {
+                sats_in += input.txin_value;
+
+                if sats_out < sats_in {
+                    traversals_cache.insert((ordinal_block_number, tx_cursor.0), lazy_tx.clone());
+                    ordinal_offset = sats_out - (sats_in - input.txin_value);
+                    ordinal_block_number = input.block_height;
+                    tx_cursor = (input.txin.clone(), input.vout as usize);
+                    break;
+                }
             }
 
-            if let Some(entry) = known_transactions.remove(&key) {
-                l1_cache_hits.push(key.clone());
-                cache_l1.insert(key, entry);
-                continue;
+            if sats_in == 0 {
+                ctx.try_log(|logger| {
+                    error!(
+                        logger,
+                        "Transaction {} is originating from a non spending transaction",
+                        transaction_identifier.hash
+                    )
+                });
+                return Ok(TraversalResult {
+                    inscription_number: 0,
+                    ordinal_number: 0,
+                    transfers: 0,
+                    inscription_input_index,
+                    transaction_identifier_inscription: transaction_identifier.clone(),
+                    transfer_data: TransferData {
+                        inscription_offset_intra_output,
+                        transaction_identifier_location: transaction_identifier.clone(),
+                        output_index: inscription_output_index,
+                        tx_index: 0,
+                    },
+                });
             }
-
-            // Enqueue for traversals
-            transactions_ids.push((
-                tx.transaction_identifier.clone(),
-                inscription_data.inscription_input_index,
-            ));
         }
     }
-    (transactions_ids, l1_cache_hits)
+
+    let height = Height(ordinal_block_number.into());
+    let ordinal_number = height.starting_sat().0 + ordinal_offset + inscription_offset_intra_output;
+
+    Ok(TraversalResult {
+        inscription_number,
+        ordinal_number,
+        transfers: hops,
+        inscription_input_index,
+        transaction_identifier_inscription: transaction_identifier.clone(),
+        transfer_data: TransferData {
+            inscription_offset_intra_output,
+            transaction_identifier_location: transaction_identifier.clone(),
+            output_index: inscription_output_index,
+            tx_index: 0,
+        },
+    })
 }
